requirements.txt
streamlit
requests
pillow
torch
torchvision
